+++ spec/models/contact_spec.rb
@@ -22,4 +22,26 @@ RSpec.describe Contact, type: :model do
       it { is_expected.to_not allow_value("notavalidemail").for(:email) }
     end
   end
+
+  describe "scopes" do
+    describe "grouped_by_category" do
+      it "groups and orders by category" do
+        mock_successful_api_responses(urn: any_args, ukprn: any_args)
+        project = create(:voluntary_conversion_project)
+        first_solicitor_contact = create(:contact, category: :solicitor, name: "B solicitor", project: project)
+        second_solicitor_contact = create(:contact, category: :solicitor, name: "A solicitor", project: project)
+        _other_contact = create(:contact, category: :other, project: project)
+
+        contacts = project.contacts.grouped_by_category
+        groups = contacts.keys
+        solicitor_group = contacts["solicitor"]
+
+        expect(groups.first).to eql "solicitor"
+        expect(groups.last).to eql "other"
+
+        expect(solicitor_group.first).to eq second_solicitor_contact
+        expect(solicitor_group.last).to eq first_solicitor_contact
+      end
+    end
+  end
 end