+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Users will use the new task list and task backend
 - Solicitor has been added to the contact category options
 - Users cannot attempt to submit the 'choose category' option on a new contact
+- The order of contacts has been updated, it is now by category and name
 
 ### Fixed